Updated Practice 5

+++ Practica5/FileUtils.h
@@ -2,5 +2,5 @@
 
 void* OpenFile(const char* _sFileName, const char* _sMode);
 bool CloseFile(void* _pFile);
-unsigned int ReadFile(char* _pBuffer, unsigned int _uSize, void* _pFile);
-unsigned int WriteFile(char* _pBuffer, unsigned int _uSize, void* _pFile);
+size_t ReadFile(char* _pBuffer, unsigned int _uSize, void* _pFile);
+size_t WriteFile(char* _pBuffer, unsigned int _uSize, void* _pFile);
